refactor: extract funcs from swiftgen.ts → extract/
