+++ src/extract/typeAliases.ts
@@ -1,0 +1,11 @@
+import { convertType } from "../utils/typeMap";
+
+export function extractTypeAliases(sourceFile: any) {
+  return sourceFile.getTypeAliases().map((alias: any) => ({
+    name: alias.getName(),
+    properties: alias.getTypeNode().getType().getProperties().map((prop: any) => ({
+      name: prop.getName(),
+      type: convertType(prop.getTypeAtLocation(prop.getDeclarations()[0]).getText())
+    }))
+  }));
+}
