+++ src/extract/variables.ts
@@ -1,0 +1,4 @@
+export function extractVariables(sourceFile: any) {
+  return sourceFile.getVariableStatements()
+    .flatMap((statement: any) => statement.getDeclarations().map((decl: any) => decl.getName()));
+}
